pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.13.0",
]
